examples(openai-agents): nest support-desk demo + fix embedding span trace
Move the autonomous support desk demo under examples/openai-agents-sdk/
and fix the search_kb embedding orphaning into its own trace.

The embedding was traced by OpenAIInstrumentor off the OTEL context,
where there's no live parent during the agent run, so it landed in a
separate single-span trace instead of under the ticket. Wrap
embeddings.create() in an Agents-SDK custom_span so it joins the same
trace, and drop OpenAIInstrumentor (plus its now-dead Responses-wrapper
workaround and the respan-instrumentation-openai dependency). The
embedding span is preserved, so the demo still showcases all five
services.
