+++ python-sdks/examples/openai-agents-sdk/autonomous-support-desk/requirements.txt
@@ -3,7 +3,6 @@
 
 respan-ai>=4.1.0                                # the `respan` facade (Respan, propagate_attributes)
 respan-sdk>=2.6.0                               # ensure_trace_id + shared types
-respan-instrumentation-openai>=1.0.1            # embedding span (search_kb)
 respan-instrumentation-openai-agents>=1.1.1     # Agents SDK tracing - MUST include the nested-trace/cost fix
 openai-agents>=0.17.6                           # the `agents` package
 openai>=2.0.0
