+++ python-sdks/examples/openai-agents-sdk/autonomous-support-desk/run_stream.py
