+++ python-sdks/examples/openai-agents-sdk/autonomous-support-desk/seed_prompts.py
